Allow scrolling to bottom of popup from overflow menu
When in overflow menu and the height of the screen is too small to show
the full popup, a scrollbar appears on overflow-y. However some part of
the popup is clipped by the box that displays it, which means that even
when scrolling to the maximum, a portion of the popup is still hidden.

As a side note, there seem to be contradictory max-heights in the XUL
elements containing the popup.

As no amount of playing with {min-,max-,}height properties on html nor
body managed to fix this, we add 50px of whitespace padding at the
bottom of the popup, which means the lost part of the popup is always
irrelevant. Only add this for documents of size <= 425px, which seems to
correspond to the overflow menu's width. Should fix #8.