Cleanup regex matching
Use unpacking instead of weird replace-based or non-standard RegExp.$n
syntax. Move 'javascript:' matching earlier as it is orthogonal to the
other tests which now can apply to the results of that match.

Also allow characters in-between quote and https?://, meaning we
identify parameters to 'javascript:' functions that *contain* absolute
URLs instead of parameters that are (or start with) absolute URLs.
Should fix #14.