Add pageAction, specifically for Android: see #30
Android does not have browserAction, so we need pageAction. However I
could not figure out a way to fully disable pageAction on desktop
firefox. If it only shows up in the manifest, it still appears as greyed
out in the page action list, and the manifest does not seem to be
adjustable on a per-platform basis.

Hence, we enable it in all platforms. Users may hide or display it.

Also add a title to the browser action popup.

TODO: The browser popup needs to be adapted in Android, as it opens in a
new tab. It either needs to figure out from which tab it was opened and
show that tab's history, or needs to show all the tabs histories
together (possibly with some hint/filtering as to in which tab this
cleaning happened).