Make httpomr preference affect UX not addon intrinsics
Instead of having the user choose which mechanism to enable, have the
mechanism always enabled and make the choice about which pages are
targeted. With the preference off, we'll only clean URLs of top-level
documents, not iframes or resources, and with it on we target all
http(s) traffic.  Closes #62