Refactor slightly test framework, move base up to root
Actually a more natural fit than using addon/ as the root since karma
always requires the /base/ prefix, so /req did not translate to addon/req.

Use the getURL() to prepend /base/addon/ for all requests except
CSS-based icon requests (are those needed?) for which we still use a proxy.